requirements.txt
azure-core==1.30.1
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
packaging==24.0
portalocker==2.8.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
